requirements.txt
Flask
firebase-admin
pandas
apscheduler
selenium
python-dotenv
